# docs/swagger.yaml
basePath: /
definitions:
...
  api.CompoundInterestRequest:
    properties:
      annualInterestRate:
        type: number
      compoundingFrequency:
        description: '"yearly" | "half_yearly" | "quarterly" | "monthly" | "daily"'
        type: string
      principal:
        type: number
      time:
        type: number
      timeUnit:
        description: '"years" | "months"'
        type: string
    type: object
...
  api.SimpleInterestRequest:
    properties:
      annualInterestRate:
        type: number
      principal:
        type: number
      time:
        type: number
      timeUnit:
        description: '"years" | "months" | "days"'
        type: string
    type: object
...
  converter.CompoundInterestResponse:
    properties:
      annualInterestRate:
        type: number
      breakdown:
        $ref: '#/definitions/converter.InterestBreakdown'
      compoundingFrequency:
        type: string
      compoundsPerYear:
        type: integer
      effectiveAnnualRate:
        type: number
      interest:
        type: number
      principal:
        type: number
      steps:
        items:
          type: string
        type: array
      timeYears:
        type: number
      totalAmount:
        type: number
      yearlyBreakdown:
        items:
          $ref: '#/definitions/converter.YearlyRow'
        type: array
    type: object
...
  converter.InterestBreakdown:
    properties:
      interestPercent:
        type: number
      principalPercent:
        type: number
    type: object
...
  converter.SimpleInterestResponse:
    properties:
      annualInterestRate:
        type: number
      breakdown:
        $ref: '#/definitions/converter.InterestBreakdown'
      interest:
        type: number
      principal:
        type: number
      steps:
        items:
          type: string
        type: array
      timeYears:
        type: number
      totalAmount:
        type: number
    type: object
...
  converter.YearlyRow:
    properties:
      closingBalance:
        type: number
      interestEarned:
        type: number
      openingBalance:
        type: number
      year:
        type: integer
    type: object
...
  nifty.CompaniesResponse:
    properties:
      companies:
        items:
          $ref: '#/definitions/nifty.Company'
        type: array
      count:
        type: integer
      error:
        type: string
      success:
        type: boolean
      updated_at:
        type: string
    type: object
  nifty.Company:
    properties:
      company_name:
        type: string
      industry:
        type: string
      isin:
        type: string
      series:
        type: string
      symbol:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Utility Microservice for converting files and user authentication.
  title: File Converter API
  version: "1.0"
paths:
...
  /calculate-compound-interest:
    post:
      consumes:
      - application/json
      description: |-
        Calculates the compound interest, total amount, effective annual rate,
        and (for tenures вүҘ 1 year) a year-by-year growth schedule.

        **Formula:** `A = P Г— (1 + r/n)^(nГ—t)`   and   `CI = A вҲ’ P`
        where `r` is the annual rate in decimal form, `n` is the number of
        compounding periods per year, and `t` is the tenure in years.

        **Limits:**
        - `principal` must be positive and вүӨ 1 trillion.
        - `annualInterestRate` must be between 0 and 100.
        - `time` must be positive; tenure must be вүӨ 100 years.
        - `timeUnit` must be `"years"` or `"months"` (day-based tenure is rejected).
        - `compoundingFrequency` must be one of `yearly`, `half_yearly`,
        `quarterly`, `monthly`, `daily` (defaults to `yearly`).
      parameters:
      - description: Compound Interest Calculation Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.CompoundInterestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Compound Interest Result
          schema:
            $ref: '#/definitions/converter.CompoundInterestResponse'
        "400":
          description: Bad Request - invalid input
          schema:
            type: string
      summary: Compound Interest Calculator
      tags:
      - Calculators
...
  /calculate-simple-interest:
    post:
      consumes:
      - application/json
      description: |-
        Calculates the simple interest and total amount for a loan or investment.

        **Formula:** `SI = (P Г— R Г— T) / 100`
        where `P` is the principal, `R` is the annual rate in percent,
        and `T` is the tenure expressed in years.

        **Limits:**
        - `principal` must be positive and вүӨ 1 trillion.
        - `annualInterestRate` must be between 0 and 100.
        - `time` must be positive; tenure must be вүӨ 100 years.
        - `timeUnit` must be `"years"`, `"months"`, or `"days"`.
      parameters:
      - description: Simple Interest Calculation Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.SimpleInterestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Simple Interest Result
          schema:
            $ref: '#/definitions/converter.SimpleInterestResponse'
        "400":
          description: Bad Request - invalid input
          schema:
            type: string
      summary: Simple Interest Calculator
      tags:
      - Calculators
...
  /nifty50/companies:
    get:
      consumes:
      - application/json
      description: Returns the latest official 50 NIFTY 50 constituent companies.
        Requires JWT Authentication.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/nifty.CompaniesResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        "503":
          description: NIFTY 50 data is currently unavailable
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get NIFTY 50 Constituent Companies
      tags:
      - markets
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"